move gaia funcs to gaia.go
